requirements.txt
mysql-connector-python==8.1.0
PyJWT==2.8.0
requests==2.32.0
python-dotenv==1.0.0